metering: expose io/cpu/rss under features=["metering"]
Move the object-store ledger into runtime_metrics::io alongside cpu and
rss so dependents can pull every in-engine signal without test-helpers.
Storage providers only record into the ledger; one ownership home.
