+++ src/lib.rs
@@ -146,9 +146,11 @@ pub use roaring;
 mod catalog;
 mod error;
 mod runtime_bridge;
-#[cfg(feature = "test-helpers")]
+// Process CPU / RSS samplers: benches use `test-helpers`; platform
+// Prometheus uses `metering`. Same module either way — no second copy.
+#[cfg(any(feature = "test-helpers", feature = "metering"))]
 pub mod runtime_metrics;
-#[cfg(not(feature = "test-helpers"))]
+#[cfg(not(any(feature = "test-helpers", feature = "metering")))]
 pub(crate) mod runtime_metrics;
 mod utils;
